requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==1.16.0
Flask==2.0.2
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.0
python-dotenv==0.19.1
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
